requirements.txt
config==0.5.1
h5py==3.14.0
imageio==2.16.0
matplotlib==3.3.4
modules==1.0.0
numpy==1.19.5
open3d==0.13.0
opencv_python==4.5.1.48
pandas==1.1.5
pcl==0.0.0.post1
pclpy==0.11.0
Pillow==11.3.0
psutil==5.9.5
python_pcl==0.3
requests==2.27.1
scikit_learn==1.7.1
scipy==1.5.4
spacepy==0.7.0
tools==1.0.5
